requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
sentence-transformers
torch
pypdf
